+++ myems-admin/app/controllers/settings/gateway/gateway.controller.js
@@ -235,6 +235,21 @@ app.controller('GatewayController', function($scope,
 		$rootScope.modalInstance = modalInstance;
 	};
 
+    $scope.copyToClipboardGateway = function (gateway) {
+        let tempInput = document.createElement("input");
+        tempInput.value = gateway.token;
+        document.body.appendChild(tempInput);
+        tempInput.select();
+        document.execCommand("copy");
+        document.body.removeChild(tempInput);
+        toaster.pop({
+            type: "success",
+            title: $translate.instant("TOASTER.SUCCESS_TITLE"),
+            body: $translate.instant("TOASTER.COPY_SUCCESS"), 
+            showCloseButton: true,
+        });         
+    };
+
 	$scope.getAllGateways();
 });