+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added tenant type function in myems-admin
 - added svg preview to myems-admin
+- added copy token function to gateway in myems-admin
 ### Changed
 - updated dashboard in myems-web and in myems-api
 - added line breaks between button and table in myems-admin